🐛 Fix: 반응형에 따라, 아이콘이 변하지 않는 문제 해결
